Lazy load the start command
